// packages/sdk/src/client/api.ts
export interface BigqueryConnection {
    /**
     * 
     * @type {string}
     * @memberof BigqueryConnection
     */
    'defaultProjectId'?: string;
    /**
     * 
     * @type {string}
     * @memberof BigqueryConnection
     */
    'billingProjectId'?: string;
    /**
     * 
     * @type {string}
     * @memberof BigqueryConnection
     */
    'location'?: string;
    /**
     * 
     * @type {string}
     * @memberof BigqueryConnection
     */
    'serviceAccountKeyJson'?: string;
    /**
     * 
     * @type {string}
     * @memberof BigqueryConnection
     */
    'maximumBytesBilled'?: string;
    /**
     * 
     * @type {string}
     * @memberof BigqueryConnection
     */
    'queryTimeoutMilliseconds'?: string;
}
// ...
export const ConnectionTypeEnum = {
    Postgres: 'postgres',
    Bigquery: 'bigquery',
    Snowflake: 'snowflake',
    Trino: 'trino'
} as const;

export type ConnectionTypeEnum = typeof ConnectionTypeEnum[keyof typeof ConnectionTypeEnum];
// ...
export interface PostgresConnection {
    /**
     * 
     * @type {string}
     * @memberof PostgresConnection
     */
    'host'?: string;
    /**
     * 
     * @type {number}
     * @memberof PostgresConnection
     */
    'port'?: number;
    /**
     * 
     * @type {string}
     * @memberof PostgresConnection
     */
    'databaseName'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostgresConnection
     */
    'userName'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostgresConnection
     */
    'password'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostgresConnection
     */
    'connectionString'?: string;
}
// ...
export interface SnowflakeConnection {
    /**
     * 
     * @type {string}
     * @memberof SnowflakeConnection
     */
    'account'?: string;
    /**
     * 
     * @type {string}
     * @memberof SnowflakeConnection
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof SnowflakeConnection
     */
    'password'?: string;
    /**
     * 
     * @type {string}
     * @memberof SnowflakeConnection
     */
    'warehouse'?: string;
    /**
     * 
     * @type {string}
     * @memberof SnowflakeConnection
     */
    'database'?: string;
    /**
     * 
     * @type {string}
     * @memberof SnowflakeConnection
     */
    'schema'?: string;
    /**
     * 
     * @type {number}
     * @memberof SnowflakeConnection
     */
    'responseTimeoutMilliseconds'?: number;
}
/**
 * 
 * @export
 * @interface TrinoConnection
 */
export interface TrinoConnection {
    /**
     * 
     * @type {string}
     * @memberof TrinoConnection
     */
    'server'?: string;
    /**
     * 
     * @type {string}
     * @memberof TrinoConnection
     */
    'port'?: string;
    /**
     * 
     * @type {string}
     * @memberof TrinoConnection
     */
    'catalog'?: string;
    /**
     * 
     * @type {string}
     * @memberof TrinoConnection
     */
    'schema'?: string;
    /**
     * 
     * @type {string}
     * @memberof TrinoConnection
     */
    'user'?: string;
    /**
     * 
     * @type {string}
     * @memberof TrinoConnection
     */
    'password'?: string;
}